docker/requirements.txt
docopt==0.6.2
pandas==1.0.0
PyYAML==5.3
schema==0.7.1
scikit-learn==0.22.1
tensorboard==2.1.0
transformers==2.3.0
